+++ src/regressors/src/DynamicSample.cpp
@@ -11,7 +11,7 @@
 
 namespace KDL {
 namespace CoDyCo {
-namespace Regressors {   
+namespace Regressors {
 
 DynamicStateSample::DynamicStateSample(const int NrOfDOFs, const double ts): n_dof(NrOfDOFs),
                                                                              timestamp(ts),
@@ -19,7 +19,7 @@ DynamicStateSample::DynamicStateSample(const int NrOfDOFs, const double ts): n_d
                                                                              dq(NrOfDOFs),
                                                                              ddq(NrOfDOFs)
 {
-    
+
 }
 
 DynamicStateSample::DynamicStateSample(const DynamicStateSample & sample):  n_dof(sample.getNrOfDOFs()),
@@ -35,7 +35,7 @@ DynamicStateSample::DynamicStateSample(const DynamicStateSample & sample):  n_do
     setJointVelocity(sample.getJointVelocity());
     setJointAcceleration(sample.getJointAcceleration());
 }
-    
+
 DynamicStateSample::~DynamicStateSample()
 {
 }
@@ -55,7 +55,7 @@ int DynamicStateSample::getNrOfDOFs() const
 {
     return n_dof;
 }
-    
+
 KDL::Frame DynamicStateSample::getWorldBaseTransform() const
 {
     return world_base_transform;
@@ -66,8 +66,8 @@ bool DynamicStateSample::setWorldBaseTransform(const KDL::Frame & _world_base_fr
     world_base_transform = _world_base_frame;
     return true;
 }
-    
-    
+
+
 KDL::Twist DynamicStateSample::getBaseVelocity() const
 {
     return base_vel;
@@ -94,37 +94,37 @@ bool DynamicStateSample::setBaseSpatialAcceleration(const KDL::Twist & _base_acc
 KDL::Twist DynamicStateSample::getBaseClassicalAcceleration() const
 {
     /**
-     * If you have doubt of this formula, check 
+     * If you have doubt of this formula, check
      * Featherstone - Rigid Body Dynamics Algorithms - 2008 - Section 2.11
-     * 
+     *
      */
     KDL::Twist classical_base_acc;
     classical_base_acc.rot = base_acc.rot;
-    classical_base_acc.vel = base_acc.vel-base_vel.vel*base_vel.rot; 
+    classical_base_acc.vel = base_acc.vel-base_vel.vel*base_vel.rot;
     return classical_base_acc;
 }
-    
+
 bool DynamicStateSample::setBaseClassicalAcceleration(const KDL::Twist & classical_base_acc)
 {
     /**
-     * If you have doubt of this formula, check 
+     * If you have doubt of this formula, check
      * Featherstone - Rigid Body Dynamics Algorithms - 2008 - Section 2.11
      */
     base_acc.rot = classical_base_acc.rot;
-    base_acc.vel = classical_base_acc.vel + base_vel.vel*base_vel.rot; 
+    base_acc.vel = classical_base_acc.vel + base_vel.vel*base_vel.rot;
     return true;
 }
-    
+
 const KDL::JntArray & DynamicStateSample::getJointPosition() const
 {
     return q;
 }
-    
+
 const KDL::JntArray & DynamicStateSample::getJointVelocity() const
 {
     return dq;
 }
-    
+
 const KDL::JntArray & DynamicStateSample::getJointAcceleration() const
 {
     return ddq;
@@ -132,7 +132,7 @@ const KDL::JntArray & DynamicStateSample::getJointAcceleration() const
 
 bool DynamicStateSample::setJointPosition(const KDL::JntArray & _q)
 {
-    if( _q.rows() != getNrOfDOFs() ) { return false; };
+    if( _q.rows() != (int)getNrOfDOFs() ) { return false; };
     q = _q;
     return true;
 }
@@ -144,25 +144,25 @@ bool DynamicStateSample::setJointPosition(double _q, int dof_id)
     return true;
 }
 
-    
+
 bool DynamicStateSample::setJointVelocity(const KDL::JntArray & _dq)
 {
-    if( _dq.rows() != getNrOfDOFs() ) { return false; };
+    if( _dq.rows() != (int)getNrOfDOFs() ) { return false; };
     dq = _dq;
     return true;
 }
 
 bool DynamicStateSample::setJointVelocity(double _dq, int dof_id)
 {
-    if( dof_id < 0 || dof_id >= getNrOfDOFs() ) { return false; }
+    if( dof_id < 0 || dof_id >= (int)getNrOfDOFs() ) { return false; }
     dq(dof_id) = _dq;
     return true;
 }
 
-    
+
 bool DynamicStateSample::setJointAcceleration(const KDL::JntArray & _ddq)
 {
-    if( _ddq.rows() != getNrOfDOFs() ) { return false; };
+    if( _ddq.rows() != (int)getNrOfDOFs() ) { return false; };
     ddq = _ddq;
     return true;
 }
@@ -174,9 +174,9 @@ bool DynamicStateSample::setJointAcceleration(double _ddq, int dof_id)
     return true;
 }
 
-    
 
-DynamicSensorSample::DynamicSensorSample(const int NrOfTorqueSensors, 
+
+DynamicSensorSample::DynamicSensorSample(const int NrOfTorqueSensors,
                                          const int NrOfWrenchSensors,
                                          const int NrOfThreeAxisSensors
                                         ):
@@ -197,32 +197,32 @@ DynamicSensorSample::DynamicSensorSample(const DynamicSensorSample & sample):
 {
     wrench_sensors_measures.resize(sample.getNrOfWrenchSensors());
     three_axis_ft_sensors_measures.resize(sample.getNrOfThreeAxisForceTorqueSensors());
-    
+
     for(int i=0; i < getNrOfTorqueSensors(); i++ ) {
         setTorqueMeasure(sample.getTorqueMeasure(i),i);
     }
-    
+
     for(int i=0; i < getNrOfWrenchSensors(); i++ ) {
         setWrenchMeasure(sample.getWrenchMeasure(i),i);
     }
-    
+
     for(int i=0; i < getNrOfThreeAxisForceTorqueSensors(); i++ ) {
         setThreeAxisForceTorqueMeasure(sample.getThreeAxisForceTorqueMeasure(i),i);
     }
-    
+
 }
 
 
-    
+
 DynamicSensorSample::~DynamicSensorSample()
 {
 }
-    
+
 int DynamicSensorSample::getNrOfWrenchSensors() const
 {
     return nrOfWrenchSensors;
 }
-    
+
 int DynamicSensorSample::getNrOfTorqueSensors() const
 {
     return nrOfTorqueSensors;
@@ -232,27 +232,27 @@ int DynamicSensorSample::getNrOfThreeAxisForceTorqueSensors() const
 {
     return nrOfThreeAxisFTSensors;
 }
-   
+
 bool DynamicSensorSample::setTorqueMeasure(const double measured_torque, const int torque_measured_id)
 {
     if( torque_measured_id < 0 || torque_measured_id >= nrOfTorqueSensors ) { return false; }
     torque_sensors_measures(torque_measured_id) = measured_torque;
     return true;
 }
-    
+
 double DynamicSensorSample::getTorqueMeasure(const int torque_measured_id) const
 {
     if( torque_measured_id < 0 || torque_measured_id >= nrOfTorqueSensors ) { return 0; }
     return torque_sensors_measures(torque_measured_id);
 }
-    
+
 bool DynamicSensorSample::setWrenchMeasure(const KDL::Wrench measured_wrench, const int wrench_measured_id)
 {
     if( wrench_measured_id < 0 || wrench_measured_id >= nrOfWrenchSensors ) { return false; }
     wrench_sensors_measures[wrench_measured_id] = measured_wrench;
     return true;
 }
-    
+
 KDL::Wrench DynamicSensorSample::getWrenchMeasure(const int wrench_measured_id) const
 {
     if( wrench_measured_id < 0 || wrench_measured_id >= nrOfWrenchSensors ) { return KDL::Wrench::Zero(); }
@@ -273,7 +273,7 @@ KDL::Vector DynamicSensorSample::getThreeAxisForceTorqueMeasure(const int three_
 }
 
 DynamicSample::DynamicSample(const int NrOfDOFs,
-                             const int NrOfTorqueSensors, 
+                             const int NrOfTorqueSensors,
                              const int NrOfWrenchSensors,
                              const int NrOfThreeAxisSensors):
                              DynamicStateSample(NrOfDOFs),
@@ -286,7 +286,7 @@ DynamicSample::DynamicSample(const DynamicSample & sample):
                              DynamicSensorSample(sample)
 {
 }
-    
+
 DynamicSample::~DynamicSample()
 {
 }
